[CHORE] redis 싱글톤 재사용 및, 리포트 실패 시에도, 메시지 생성
